feat: ヘッダーの処理をhomeに移す
